requirements.txt
# ================================================================
# DEPENDENCIAS DEL ETL FRESHSALE -> SQL SERVER
# ================================================================

# Cliente HTTP
requests>=2.31.0

# SQL Server driver
pymssql>=2.3.0

# Variables de entorno
python-dotenv>=1.0.0

# Parser de fechas
python-dateutil>=2.8.2
